WIP: Feature to enable uniform behavior for unix, windows
This patch tries to gives Unix the same behavior as is documented for
Windows when opening multiple handles to the same file: if any other
Lockfile currently is locking the same file, then other Lockfiles
will block when trying to lock it.

It works by using an internal HashMap to determine which files are
locked and which are not.  (We identify files by a combination of
st_dev and st_ino).  This logic is behind a new feature,
`multilock`, which depends on `std` and `once_cell`.

I'm submitting this patch for feedback, to ask if the general
approach looks like something that could be accepted. Please don't
merge it as-is: it needs tests and documentation.