Refactor to avoid feature-dependent behavior for `LockFile::open`
It's a bad idea to have the same function behave differently
depending on whether a feature is enabled: If one crate depends on
the behavior without the feature, and another crate enables the
feature, then those two crates can't be linked together without
breaking the first one.

So this patch changes `multilock` so that it never changes the
behavior of `LockFile::open`.  Instead, there is a new
`LockFile::open_excl` call that is available when `multilock` is
enabled on Unix, and always on Windows.